requirements.txt
annotated-types==0.7.0
anyio==4.8.0
automata-lib==9.0.0
cached_method==0.1.0
click==8.1.8
colorama==0.4.6
fastapi==0.115.8
frozendict==2.4.6
graphviz==0.20.3
h11==0.14.0
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
networkx==3.4.2
pydantic==2.10.6
pydantic_core==2.27.2
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
